Remove e2e from requirements to publish (#199)
e2e are notoriously flakey, and are just getting in our way when deploying. We've started moving to Cypress which seems to be doing much better. Our e2e should still run, so we can manually check to see if failures are legit, but otherwise, once merged, deployment will happen without their passing.