Handle issues with city data causing errors
Some events ended up with data like:
`{"nameWithHierarchy":"City Name","$$hashKey":"object:1234"}`
in the city field. The hashKey prop appears to be coming from angular.

This was breaking editing the events as the prop key was not allowed but
was submitted with the form.

We now pull out the `nameWithHierarchy` prop if it is there and strip
the rest.

We also handle the case where the dojo city value is null or empty when
creating your first event.