Create a wrapper component to set the text direction for i18n
If direction is passed in as a prop, this is used, otherwise the component
will check the store for the chosen language config and use the dir property
or default to 'ltr';
Add methods / properties to the store with actions and mutators.
When the language picker is created, update the store with the language config
matching the one chosen by the user / cookie.